+++ packages/fiori-app-sub-generator/src/fiori-app-generator/prompting.ts
@@ -4,6 +4,7 @@ import type { InputQuestion } from '@sap-ux/inquirer-common';
 import type { Logger } from '@sap-ux/logger';
 import {
     type CapService,
+    type ConnectedSystem,
     DatasourceType,
     type OdataServicePromptOptions,
     OdataVersion,
@@ -172,12 +173,14 @@ export async function promptUI5ApplicationAnswers(
  * @param options
  * @param logger
  * @param adapter
+ * @param connectedSystem
  * @returns {Promise<Service>}
  */
 export async function promptOdataServiceAnswers(
     options: OdataServiceInquirerOptions,
     logger: Logger,
-    adapter: Adapter
+    adapter: Adapter,
+    connectedSystem?: ConnectedSystem
 ): Promise<Service> {
     let inquirerAdapter;
     // type `any` will be replaced when we can import ESM modules
@@ -196,7 +199,8 @@ export async function promptOdataServiceAnswers(
         logger,
         isFeatureEnabled(Features.enableGAIntegration),
         ClientFactory.getTelemetryClient(),
-        getHostEnvironment() !== hostEnvironment.cli
+        getHostEnvironment() !== hostEnvironment.cli,
+        connectedSystem
     );
 
     const service: Service = {
@@ -210,7 +214,7 @@ export async function promptOdataServiceAnswers(
         source: answers.datasourceType,
         localEdmxFilePath: answers.metadataFilePath,
         connectedSystem: answers.connectedSystem,
-        ignoreCertError: (answers as any).ignoreCertError // TBI: ignoreCertError is not a property OdataServiceAnswers
+        ignoreCertError: answers.ignoreCertError
     };
     return service;
 }
@@ -300,6 +304,9 @@ export async function createUI5ApplicationPromptOptions(
             },
             [ui5AppInquirerPromptNames.enableTypeScript]: {
                 defaultValue: defaultPromptValues[ui5AppInquirerPromptNames.enableTypeScript]
+            },
+            [ui5AppInquirerPromptNames.enableVirtualEndpoints]: {
+                hide: service.capService?.capType === 'Java'
             }
         } as UI5ApplicationPromptOptions,
         promptSettings as UI5ApplicationPromptOptions