+++ packages/preview-middleware-client/types/sap.ui.fl.d.ts
@@ -83,7 +83,7 @@ declare module 'sap/ui/fl/Scenario' {
 declare module 'sap/ui/fl/Utils' {
     import type ManagedObject from 'sap/ui/base/ManagedObject';
     import type Controller from 'sap/ui/core/mvc/Controller';
-    import type Component from 'sap/ui/core/Component'
+    import type Component from 'sap/ui/core/Component';
 
     interface Utils {
         checkControlId(control: ManagedObject): boolean;
@@ -95,6 +95,7 @@ declare module 'sap/ui/fl/Utils' {
     interface ControlView {
         getId(): string;
         getController(): Controller;
+        getControllerModuleName?: () => string;
     }
 
     const Utils: Utils;
@@ -119,7 +120,12 @@ declare module 'sap/ui/fl/write/api/connectors/ObjectStorageConnector' {
         getItem(key: string): unknown;
         getItems(): Promise<unknown[]>;
         fileChangeRequestNotifier:
-            | (<T extends object, U extends object>(fileName: string, kind: 'create' | 'delete', change?: T, additionalChangeInfo?: U) => void)
+            | (<T extends object, U extends object>(
+                  fileName: string,
+                  kind: 'create' | 'delete',
+                  change?: T,
+                  additionalChangeInfo?: U
+              ) => void)
             | undefined;
     }