+++ packages/ui5-application-inquirer/test/unit/prompts/prompts.test.ts
@@ -440,6 +440,13 @@ describe('getQuestions', () => {
             (question) => question.name === promptNames.enableVirtualEndpoints
         );
         expect((enableVirtualEndpointsQuestion?.when as Function)({ enableTypeScript: true })).toBe(true);
+
+        // CAP project with cds-ui5 plugin disabled and hasMinCdsVersion is false
+        questions = getQuestions([], {}, { ...mockCdsInfo, isCdsUi5PluginEnabled: false, hasMinCdsVersion: false });
+        enableVirtualEndpointsQuestion = questions.find(
+            (question) => question.name === promptNames.enableVirtualEndpoints
+        );
+        expect((enableVirtualEndpointsQuestion?.when as Function)()).toBe(false);
     });
 
     test('getQuestions, prompt: `ui5Theme`', async () => {
@@ -553,6 +560,7 @@ describe('getQuestions', () => {
         });
         enableTypeScriptQuestion = questions.find((question) => question.name === promptNames.enableTypeScript);
         expect(enableTypeScriptQuestion?.default()).toEqual(true);
+        expect(enableTypeScriptQuestion?.additionalMessages!(true)).toEqual(undefined);
 
         // when
         expect((enableTypeScriptQuestion?.when as Function)()).toEqual(true);
@@ -567,10 +575,6 @@ describe('getQuestions', () => {
         );
         expect((enableTypeScriptQuestion?.when as Function)()).toEqual(false);
 
-        enableTypeScriptQuestion = getQuestions([], undefined, mockCdsInfoFalse).find(
-            (question) => question.name === promptNames.enableTypeScript
-        );
-
         enableTypeScriptQuestion = getQuestions([], undefined, {
             hasCdsUi5Plugin: true,
             isCdsUi5PluginEnabled: true,
@@ -578,6 +582,7 @@ describe('getQuestions', () => {
             hasMinCdsVersion: true
         }).find((question) => question.name === promptNames.enableTypeScript);
         expect((enableTypeScriptQuestion?.when as Function)()).toEqual(true);
+        expect(enableTypeScriptQuestion?.additionalMessages!(true)).toEqual(undefined);
 
         enableTypeScriptQuestion = getQuestions([], undefined, {
             hasCdsUi5Plugin: false,
@@ -585,6 +590,7 @@ describe('getQuestions', () => {
             isWorkspaceEnabled: false,
             hasMinCdsVersion: true
         }).find((question) => question.name === promptNames.enableTypeScript);
+        expect((enableTypeScriptQuestion?.when as Function)()).toEqual(true);
         expect(enableTypeScriptQuestion?.additionalMessages!(true)).toEqual({
             message:
                 'The CAP project will be updated to use NPM workspaces (this is a requirement for generating with TypeScript)',