+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.0.88
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.28
+
+## 0.0.87
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.27
+
+## 0.0.86
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.26
+
 ## 0.0.85
 
 ### Patch Changes