+++ packages/annotation-generator/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/annotation-generator
 
+## 0.3.30
+
+### Patch Changes
+
+-   Updated dependencies [ea0e2c0]
+    -   @sap-ux/project-access@1.29.22
+    -   @sap-ux/fiori-annotation-api@0.5.2
+
+## 0.3.29
+
+### Patch Changes
+
+-   Updated dependencies [2024a9f]
+    -   @sap-ux/fiori-annotation-api@0.5.1
+
+## 0.3.28
+
+### Patch Changes
+
+-   Updated dependencies [2a16531]
+    -   @sap-ux/fiori-annotation-api@0.5.0
+
 ## 0.3.27
 
 ### Patch Changes