+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.8.43
+
+### Patch Changes
+
+-   Updated dependencies [1a01c5e]
+    -   @sap-ux/axios-extension@1.20.2
+
+## 0.8.42
+
+### Patch Changes
+
+-   Updated dependencies [a3a43b2]
+    -   @sap-ux/axios-extension@1.20.1
+
 ## 0.8.41
 
 ### Patch Changes