+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.1.59
+
+### Patch Changes
+
+-   Updated dependencies [ea0e2c0]
+    -   @sap-ux/project-access@1.29.22
+    -   @sap-ux/cf-deploy-config-writer@0.2.13
+    -   @sap-ux/fiori-generator-shared@0.11.3
+    -   @sap-ux/deploy-config-generator-shared@0.0.47
+    -   @sap-ux/inquirer-common@0.6.42
+    -   @sap-ux/cf-deploy-config-inquirer@0.2.36
+
+## 0.1.58
+
+### Patch Changes
+
+-   Updated dependencies [b339248]
+    -   @sap-ux/cf-deploy-config-writer@0.2.12
+
+## 0.1.57
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.41
+-   @sap-ux/cf-deploy-config-inquirer@0.2.35
+-   @sap-ux/deploy-config-generator-shared@0.0.46
+
+## 0.1.56
+
+### Patch Changes
+
+-   Updated dependencies [42d73c0]
+    -   @sap-ux/cf-deploy-config-writer@0.2.11
+
+## 0.1.55
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.40
+-   @sap-ux/cf-deploy-config-inquirer@0.2.34
+-   @sap-ux/deploy-config-generator-shared@0.0.46
+
+## 0.1.54
+
+### Patch Changes
+
+-   65e178f: align npm modules
+-   Updated dependencies [65e178f]
+    -   @sap-ux/deploy-config-generator-shared@0.0.46
+    -   @sap-ux/cf-deploy-config-writer@0.2.10
+
 ## 0.1.53
 
 ### Patch Changes